Only use errors.Is backwards compatibility code, delete gobackcomp pkg
