+++ server/errors.go
@@ -16,7 +16,6 @@ package server
 import (
 	"errors"
 	"fmt"
-	"github.com/nats-io/nats-server/v2/server/gobackcomp"
 )
 
 var (
@@ -248,6 +247,39 @@ func UnpackIfErrorCtx(err error) string {
 	return err.Error()
 }
 
+// implements: go 1.13 errors.Unwrap(err error) error
+// TODO replace with native code once we no longer support go1.12
+func errorsUnwrap(err error) error {
+	u, ok := err.(interface {
+		Unwrap() error
+	})
+	if !ok {
+		return nil
+	}
+	return u.Unwrap()
+}
+
+// implements: go 1.13 errors.Is(err, target error) bool
+// TODO replace with native code once we no longer support go1.12
 func IsErr(err, target error) bool {
-	return gobackcomp.ErrorsIs(err, target)
+	// this is an outright copy of go 1.13 errors.Is(err, target error) bool
+	// removed isComparable
+	if err == nil || target == nil {
+		return err == target
+	}
+
+	for {
+		if err == target {
+			return true
+		}
+		if x, ok := err.(interface{ Is(error) bool }); ok && x.Is(target) {
+			return true
+		}
+		// TODO: consider supporing target.Is(err). This would allow
+		// user-definable predicates, but also may allow for coping with sloppy
+		// APIs, thereby making it easier to get away with them.
+		if err = errorsUnwrap(err); err == nil {
+			return false
+		}
+	}
 }
